pyproject.toml
[tool.poetry]
name = "coding-tool-reasoning"
version = "0.1.0"
description = "A short description of my awesome package."
authors = ["Andrey Justo <andrey.justo@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["python", "packaging", "example"]
package-mode = false

[tool.poetry.dependencies]
agent-framework-devui = "1.0.0b251016"
agent-framework = "1.0.0b251016"
azure-identity = "1.25.1"
bert-score = "0.3.13"
logging = "0.4.9.6"
matplotlib = "3.10.7"
networkx = "3.6"
numpy = "^2.2.6"
openai = "2.8.1"
python = "<3.14,>=3.11"
pyyaml = "6.0.3"
requests = "^2.28.1"
torch = "2.6.0" # Review versionin

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^22.3.0"
flake8 = "^4.0.1"
pytest-asyncio = "1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"